finish moving components into their own folders
now to break apart things like the forms info common elements
and move the stories to each folder
