+++ src/components/Alert/index.ts
@@ -1,0 +1,2 @@
+import Alert from "./Alert";
+export default Alert;
