+++ src/components/AuthenticationForm/index.ts
@@ -1,0 +1,2 @@
+import AuthenticationForm from "./AuthenticationForm";
+export default AuthenticationForm;
