+++ src/components/BookmarkEditForm/index.ts
@@ -1,0 +1,2 @@
+import BookmarkEditForm from "./BookmarkEditForm";
+export default BookmarkEditForm;
