+++ src/components/BuildInfo/index.ts
@@ -1,0 +1,2 @@
+import BuildInfo from "./BuildInfo";
+export default BuildInfo;
