+++ src/components/Button/index.ts
@@ -1,0 +1,2 @@
+import Button from "./Button";
+export default Button;
