+++ src/components/CenteredContent/index.ts
@@ -1,0 +1,2 @@
+import CenteredContent from "./CenteredContent";
+export default CenteredContent;
