+++ src/components/DebugFilterList/index.ts
@@ -1,0 +1,2 @@
+import DebugFilterList from "./DebugFilterList";
+export default DebugFilterList;
