+++ src/components/DevOptionsForm/index.ts
@@ -1,0 +1,2 @@
+import DevOptionsForm from "./DevOptionsForm";
+export default DevOptionsForm;
