+++ src/components/ErrorBoundary/index.ts
@@ -1,0 +1,2 @@
+import ErrorBoundary from "./ErrorBoundary";
+export default ErrorBoundary;
