+++ src/components/Loader/index.ts
@@ -1,0 +1,2 @@
+import Loader from "./Loader";
+export default Loader;
