+++ src/components/PopupContainer/index.ts
@@ -1,0 +1,2 @@
+import PopupContainer from "./PopupContainer";
+export default PopupContainer;
